Discuss the Sprint Planning format
We've had pretty good success with the Sprint Planning format
suggested in https://github.com/ipfs/community/issues/28 so this
commit adds a rough description of it to this repo.

Note: the sprint discussion times are quite late for people in
GMT+2 to join. I suggest we move up the discussion up by 2 hours
for the future (next week). I'll poll the group today and see what
is best for everyone.