Remove call title from notes link
This is unnecessary, and means I need to edit the url just a bit more each time there is an edit. I think just saying notes does the job, as it is in a row for each call topic anyway, and as the URL is not opaque itself regarding to which topci it is.