Remove lead from templates, other edits
I think having a lead for each call gives the wrong impression of the scope of each project. For instance, @haadcode isnt the only person who can handle Apps on IPFS. As well, it stops others from jumping in to facilitate these calls. Finally, in at least the All hands call, it rotates, and I think it should in the others. I have removed it in this PR.

I have also removed the extra time zone markings, which are in the column, and removed bold from the UTC timezone (which did not show up in the markdown, anyway), and added the word agenda to the notes to make it clearer what that link points to. Overall, I think this is now a cleaner template for the sprints.

Finally, I moved IPLD to Friday, and added a Monday and a Friday section to make this clearer.