Change team structures based on alt role proposal
For WGs where "the Captain is performing all roles", I also added the "People Coach" role. What I mean by this designation is just documenting the existing practice of the Captain being the first-level router of support needs and resourcing to other channels (along with being the technical lead and the process herder). This doesn't mean Captains do anything more than what they're currently doing - but it does give more clarity for team members about how to communicate their needs (if/when a need arises). If any Captain feels uncomfortable with this change, it's a good opportunity to ask the question "if someone on your team has a support or interpersonal need, who is the first person they should talk to?" - whoever that person is, check with them that they are willing to be the first router for those requests and add them separately as the WG "People Coach". =]