Research Group --> Special Interest Group (#928)
* Research Group --> Special Interest Group

I propose that we reframe "research groups" to be more inclusive of other types of work in addition to research, which can be taken on by a community to improve IPFS for a specific vertical. I'm proposing the name "Special Interest Group" since it's a common term for this type of group in ACM and other open source communities (ex Kubernetes). 

The definition of the group mostly stays the same, with the caveat that a SIG can also drive activities like building tools, creating demos, or forming non-research relationships. The big distinction between a SIG and a WG is that SIGs don't have captains or dedicated OKR sheets (though may track their goals for the quarter as OKRs if useful) - and all members of a SIG also have a main Working Group they contribute to.

I also added some indexes for ongoing tiger teams =]

* Update TEAMS_ROLES_STRUCTURES.md

Co-Authored-By: momack2 <momack2@users.noreply.github.com>